+++ packages/google_maps_flutter/google_maps_flutter_ios/lib/src/google_maps_flutter_ios.dart
@@ -208,6 +208,11 @@ class GoogleMapsFlutterIOS extends GoogleMapsFlutterPlatform {
   }
 
   @override
+  Stream<MapLoadedEvent> onMapLoaded({required int mapId}) {
+    return _events(mapId).whereType<MapLoadedEvent>();
+  }
+
+  @override
   Future<void> updateMapConfiguration(
     MapConfiguration configuration, {
     required int mapId,
@@ -1126,6 +1131,11 @@ class HostMapMessageHandler implements MapsCallbackApi {
     streamController.add(
       MapTapEvent(mapId, _latLngFromPlatformLatLng(position)),
     );
+  }
+
+  @override
+  void onMapLoaded() {
+    streamController.add(MapLoadedEvent(mapId));
   }
 }